requirements.txt
langchain
Pillow
python-dotenv
streamlit
telebot
openai
pypdf
chromadb
tiktoken
tabulate
boto3